feat: v1.4.0 — 性能扩展 + 数据治理 + WSL2 代理修复
Phase 0 (补债拆分):
- DuckDB Profiling 支持 (enable_profiling 配置)
- DuckDB Relational API 试点 (metadata.py)
- 大文件拆分: ingestor 870→200行, client 838→300行

Phase 1 (性能规模):
- DuckDB 水平扩展: 多实例连接池路由 + Redis 信号量
- GPU Autoscaling: 冷却期 + 缩容保护 + 事件持久化
- Schema 演进: SchemaCompatibilityChecker + migrate 端点
- Daft 原生媒体: 批量图像/视频 + 感知哈希

Phase 2 (数据治理):
- 血缘可视化 API: graph/impact/stats 端点
- 质量规则引擎: QualityRuleEngine (4 checks/3 actions)
- 行级/列级 ACL: DatasetACL + admin 管理端点

Phase 3 (收尾加固):
- FTS 真正分页: offset 参数全链路支持
- OLAP SSE 流式: Arrow IPC batch base64 事件流

部署修复:
- WSL2 mirrored 模式代理: proxy-forwarder socat 转发器
- core/http.py 显式代理: trust_env=False + proxy= 参数
- Loki 3.0 缓存配置修复 (fifocache → memcached)
- 31 个 API 示例脚本全部验证通过

+++ tests/unit/facade/test_admin_facade.py
@@ -70,9 +70,9 @@ class TestAddColumn:
 
 class TestAlterColumn:
     def test_alter_column_type(self, populated_lake: Lake) -> None:
-        populated_lake.alter_column("test_ds", "score", pa.float32())
+        populated_lake.alter_column("test_ds", "score", pa.float64())
         result = populated_lake.read_dataset("test_ds", columns=["score"])
-        assert result.schema.field("score").type == pa.float32()
+        assert result.schema.field("score").type == pa.float64()
 
 
 class TestDropColumn:
